requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
Flask==1.1.2
Flask-Cache==0.13.1
Flask-Caching==1.8.0
Flask-Cors==3.0.8
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paramiko==2.7.1
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.3.0
python-telegram-bot==12.7
six==1.14.0
soupsieve==2.0
sshtunnel==0.1.5
tornado==6.0.4
Werkzeug==1.0.1
